Refactor/Tests: Move response header preparation from SimpleHandler.get()
Previously, the response header preparation all happened in get() but now it
happens in a separate method that's easier to test. So I also added tests. I'm
really not sure about these templated tests... it feels like it should be more
realiable and easier to adjust, but I'm just not sure.

Almost finished... just have to refactor the unit tests for SimpleHandler.get()

Related to #31