SimpleHandler creates default values for "Page" and "Per-Page" request headers
Previously, SimpleHandler would let the defaults (None) for those request headers
be decided by Solr. This could create problems, and it was allowing issue #46 to
surface (although that was really my sloppy programming). Now, we make our own
defaults for these headers, which should remove a couple of logical branches.