SimpleHandler.make_response_headers() doesn't automatically add "type" to Fields
Previously, make_response_headers() would automatically add "type" to the
X-Cantus-Fields response header, since it would always be present in every
resource, because Abbott put it there. We're moving toward a world in which
Solr provides the "type" field, so it should be added in the usual way.

Working toward #47 (believe it or not)