New: add a proper setting for Access-Control-Allow-Origin response header
Previously, the A-C-Expose-Headers and A-C-Allow-Headers response headers were
added for every request, and the A-C-Allow-Origin header was only added, by
default, if the server were in DEBUG mode. That doesn't work in production, so
I made a setting to help deal with these properly (i.e., without modifying
anything but the "settings" file on the production server).

Working toward #39
Working toward #42