New: @request_wrapper decorator for RequestHandler methods
This decorator will prevent the sort of ambiguous situation with uncaught
exceptions that's exemplified in issue #50. The wrapper function catches
any exceptions raised in the handler method, prints an error message to
the console or journal, and sends a 500 response to the user agent.
Previously, the user agent would receive no response, and the journal
would simply say "uncaught exception."

Working toward #50