Fixedish: automatic redirects for URLs not ending with /
URLs that don't end in a / are now automatically redirected to the same
URL with a / at the end. This establishes a canonical form of Cantus API
URLs while also letting the no-slash versions work.

There's one catch: the automated test for redirecting a SEARCH request
doesn't work. It works in practice---just not in the test.

Working toward #18