Change: SimpleHandler.search_handler() and ComplexHandler.s..._h...() identical
In a fun trick of subclassing, both classes can have identical search_handler()
methods BUT the get_handler() they call is different in the two classes. In the
very near future, these methods won't be identical any more, because
ComplexHandler will have to do its sub-search for cross-referenced fields
in this method.

However, this does solve the problem that xreffed fields aren't looked up
in the responses for SEARCH requests.

Working toward #48