New: HolyOrders downloads chant updates in the new way
Drupal gave a 500 on chant update URLs when "too many" chants were updated
on the same day. We're switching to a system where Drupal only provides the
IDs of the chants that were updated, and then HolyOrders must download each
chant individually.

This is *way* more requests, so I have a feeling it's going to create some
problem somewhere... but hopefully it improves the situation somewhat.