Fixed: IndexError crash in download_chant_updates()
Previously, if given config URLs that didn't have the proper format (which turns
out to have been all the time) download_chant_updates() would raise an
IndexError. Now, the function checks that the proper substitution fields are
present in the config URLs, and exits cleanly (leaving a log message) if
they're not.