Fixed: Abbot and HolyOrders run from the virtualenv
Previously, Abbot and HolyOrders ran from their git repositories. This could
lead to a situation where some modules are imported from the virtualenv and
some from the local repository. Although these should always be the same,
it's a possible source of confusion, and also a possible attack vector.

Now, everything runs from the virtualenv. This does, however, mean that any
changes I want to deploy must be provoked with a version number increase, or
pip won't think it's an update. But I should be doing that anyway.