New: shared.TestHandler puts a mock on abbot.util.SOLR
Calling public methods on this mock raises an AssertionError telling you to
fix your tests, which should never "hit" this default Solr mock. The idea
is, if you're using Solr in your test, you should have to configure it
specifically, which I'm about to make easier in the following commits.

For #81 and #82