Fixed: util.request_wrapper calls the "log" properly
Previously, request_wrapper() would write a traceback into the log as a single
string with newlines in it. However, journald stores only the first line, so
the helpful parts of the traceback were missing. Now, request_wrapper() will
call log.debug() with every line of the traceback.

I also took the opportunity to adjust the debug levels: putting decorators in
the wrong order is an ERROR, but the tracebacks are only DEBUG, so that the log
will only have tracebacks in them when the server is set to "DEBUG."