Fixed: SimpleHandler.basic_get() does not assume "id" and "type" present
Previously, basic_get() simply assumed the "id" and "type" fields are present
in every resource returned by Solr. If they weren't, the result would be a
"500 Programmer Error" to the user agent. Now, if a resource is missing one of
these two required fields, we return a 502 to the user agent, so they know
it's Solr's fault.

Closes #98