Fixed: request_wrapper decorator only prints traceback in debug mode
Previously, the request_wrapper decorator would always print a traceback
in the journald log. This opens a vector for enterprising attackers to cause
Abbot to fail if the system log becomes full. Now that tracebacks are only
printed to the log if the server is in debug mode, this becomes much more
difficult.

Closes #101