Fixed: fields in search queries are validated before submission
Previously, an invalid field in a search query would be passed along to
Solr, resulting in a 502 response to the user agent. This isn't helpful,
because it implies that Solr is having a problem, when it's really the user
who submitted an invalid query.

Thus Abbot now checks fields before submitting the query, and returns a 400
when there is an invalid field.

Closes #103