Fixed: redirects with query strings fail properly
Previously, URLs like this:
   /test?one=two
would result in infinite redirects as we tried to save the situation by
adding a / to the end of the query string, when we really wanted it just
before the query string.

Now, redirects with a query string just fail. This is a rather blunt
approach, but the Cantus API never uses query strings, so if an incoming
URL has one then something is wrong.