Change: compile Vitrail locally then send to server
Previously, Vitrail was compiled on the server, meaning NodeJS, the Vitrail
repository, and all the NPM depenencies had to be installed on the server.
It takes a lot of disk space and memory, so that running the playbook caused
swap usage and consequently affected the server's ability to run Solr/Abbot.

Now all of that stuff is done on the local machine that runs the playbook,
so the server only takes on the files and software needed for deployment.